chore: convert watch_test into an integration test
